+++ CHANGELOG.md
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - `Log::gadgets` to retrieve all gadget agents.
 - `Log::build_id` to retrieve the game's build id.
 - The `serde` optional feature to enable (de)serialization of API types.
+- `Encounter::from_header_id` to convert a header ID from arcdps to the correct
+  encounter.
 
 ### Changed
 - `gamedata::Boss` has been split in `gamedata::Boss` and `gamedata::Encounter`